provision-database.handler.test first pass
