+++ test/aurora.provisioner-user.handler.test.ts
@@ -9,7 +9,7 @@ import {
 import { Client } from 'pg';
 import sinon from 'sinon';
 
-import { Methods, handler } from '../src/aurora.provisioner';
+import { Methods, handler } from '../src/aurora.provision-user';
 
 sinon.stub(console, 'log');
 
@@ -18,8 +18,7 @@ const originalEnv = process.env;
 const fetchAndConformSecretsStub = sinon.stub(Methods.prototype, 'fetchAndConformSecrets');
 const createUserStub = sinon.stub(Methods.prototype, 'createUser');
 const conformPasswordStub = sinon.stub(Methods.prototype, 'conformPassword');
-const createDatabaseStub = sinon.stub(Methods.prototype, 'createDatabase');
-const grantPrivilegesStub = sinon.stub(Methods.prototype, 'grantPrivileges');
+const grantRoleStub = sinon.stub(Methods.prototype, 'grantRole');
 const postgresStub = sinon.stub(Client.prototype, 'connect');
 
 const standardSecretResult = {
@@ -37,8 +36,7 @@ beforeEach(() => {
   fetchAndConformSecretsStub.reset();
   createUserStub.reset();
   conformPasswordStub.reset();
-  createDatabaseStub.reset();
-  grantPrivilegesStub.reset();
+  grantRoleStub.reset();
   postgresStub.reset();
 
   jest.resetModules();
@@ -54,6 +52,7 @@ describe('handler', () => {
     ServiceToken: 'fakeServiceToken',
     userSecretArn: 'fakeUserSecretArn',
     isWriter: true,
+    proxyHost: 'fakeProxyHost',
   };
 
   const eventBase: CloudFormationCustomResourceEventCommon = {
@@ -84,11 +83,11 @@ describe('handler', () => {
   const callback: Callback = (_err, _data) => {};
 
   describe('runs onCreate', () => {
+    const createEvent: CloudFormationCustomResourceCreateEvent = {
+      ...eventBase,
+      RequestType: 'Create',
+    };
     it('no MANAGER_SECRET_ARN', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-      };
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
         LogicalResourceId: 'fakeLogicalResourceId',
@@ -101,10 +100,6 @@ describe('handler', () => {
     });
 
     it('handles error from fetchAndConformSecrets', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-      };
       process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
       fetchAndConformSecretsStub.rejects(new Error('whoopsie'));
       const r = await handler(createEvent, context, callback);
@@ -119,10 +114,6 @@ describe('handler', () => {
     });
 
     it('handles error from postgres', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-      };
       process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
       fetchAndConformSecretsStub.resolves(standardSecretResult);
       postgresStub.rejects(new Error('whoopsie'));
@@ -138,10 +129,6 @@ describe('handler', () => {
     });
 
     it('handles error from createUser', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-      };
       process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
       fetchAndConformSecretsStub.resolves(standardSecretResult);
       postgresStub.resolves();
@@ -158,10 +145,6 @@ describe('handler', () => {
     });
 
     it('handles error from conformPassword', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-      };
       process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
       fetchAndConformSecretsStub.resolves(standardSecretResult);
       postgresStub.resolves();
@@ -172,6 +155,24 @@ describe('handler', () => {
         LogicalResourceId: 'fakeLogicalResourceId',
         PhysicalResourceId: 'fakeUsername',
         Reason: 'Create / conform issue: Error: whoopsie see also fakeLogStreamName',
+        RequestId: 'fakeRequestId',
+        StackId: 'fakeStackId',
+        Status: 'FAILED',
+      });
+    });
+
+    it('handles error from grantRole', async () => {
+      process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
+      fetchAndConformSecretsStub.resolves(standardSecretResult);
+      postgresStub.resolves();
+      createUserStub.resolves();
+      conformPasswordStub.resolves();
+      grantRoleStub.rejects(new Error('whoopsie'));
+      const r = await handler(createEvent, context, callback);
+      expect(r).toEqual({
+        LogicalResourceId: 'fakeLogicalResourceId',
+        PhysicalResourceId: 'fakeUsername',
+        Reason: 'Grant role issue: Error: whoopsie see also fakeLogStreamName',
         RequestId: 'fakeRequestId',
         StackId: 'fakeStackId',
         Status: 'FAILED',
@@ -216,39 +217,12 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.rejects(new Error('whoopsie'));
+      grantRoleStub.rejects(new Error('whoopsie'));
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
         LogicalResourceId: 'fakeLogicalResourceId',
         PhysicalResourceId: 'fakeUsername',
         Reason: 'Create database issue: Error: whoopsie see also fakeLogStreamName',
-        RequestId: 'fakeRequestId',
-        StackId: 'fakeStackId',
-        Status: 'FAILED',
-      });
-    });
-
-    it('handles error from postgres re-connect', async () => {
-      const createEvent: CloudFormationCustomResourceCreateEvent = {
-        ...eventBase,
-        RequestType: 'Create',
-        ResourceProperties: {
-          ...resourcePropertiesBase,
-          databaseName: 'fakeDbName',
-        },
-      };
-      process.env.MANAGER_SECRET_ARN = 'fakeManagerSecretArn';
-      fetchAndConformSecretsStub.resolves(standardSecretResult);
-      postgresStub.onFirstCall().resolves();
-      createUserStub.resolves();
-      conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
-      postgresStub.onSecondCall().rejects(new Error('whoopsie'));
-      const r = await handler(createEvent, context, callback);
-      expect(r).toEqual({
-        LogicalResourceId: 'fakeLogicalResourceId',
-        PhysicalResourceId: 'fakeUsername',
-        Reason: 'client.connect failed: Error: whoopsie see also fakeLogStreamName',
         RequestId: 'fakeRequestId',
         StackId: 'fakeStackId',
         Status: 'FAILED',
@@ -269,7 +243,7 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
+      grantRoleStub.resolves();
       grantPrivilegesStub.rejects(new Error('whoopsie'));
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
@@ -296,7 +270,7 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
+      grantRoleStub.resolves();
       grantPrivilegesStub.resolves();
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
@@ -425,7 +399,7 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.rejects(new Error('whoopsie'));
+      grantRoleStub.rejects(new Error('whoopsie'));
       const r = await handler(updateEvent, context, callback);
       expect(r).toEqual({
         LogicalResourceId: 'fakeLogicalResourceId',
@@ -450,7 +424,7 @@ describe('handler', () => {
       postgresStub.onFirstCall().resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
+      grantRoleStub.resolves();
       postgresStub.onSecondCall().rejects(new Error('whoopsie'));
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
@@ -476,7 +450,7 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
+      grantRoleStub.resolves();
       grantPrivilegesStub.rejects(new Error('whoopsie'));
       const r = await handler(updateEvent, context, callback);
       expect(r).toEqual({
@@ -502,7 +476,7 @@ describe('handler', () => {
       postgresStub.resolves();
       createUserStub.resolves();
       conformPasswordStub.resolves();
-      createDatabaseStub.resolves();
+      grantRoleStub.resolves();
       grantPrivilegesStub.resolves();
       const r = await handler(createEvent, context, callback);
       expect(r).toEqual({
